Support current draftsharing behaviour in Wagtail 6
This commit mirrors the current functionality for Wagtail 5 in Wagtail 6 by
creating a Stimulus JS controller and hooking it into Wagtail in a similar
way as for Wagtail 5.

For the page history view, the HTML templating is different, so the previous
approach overrode wagtailadmin/pages/revisions/_actions.html to slot in custom
markup that the JS could hook into. For Wagtail 6 we do something similar,
extending wagtailadmin/generic/history/action_cell.html to slot in a bit of
custom markup.

For snippet-list view, we update the HTML output for the share_url method
to hook in the Stimulus controller used above.

The JS in the Stimulus controller is deliberately very similar to the
original JS. I wondered about refactoring out the main duplicated code
(the main function which triggers the creation of new links via a POST
to the backend, and the function which upgrades the View button on the
list view of existing sharing links to be a clipboard link). However,
for now - and with a view potentially to deprecating pre-Wagtail 6 behaviour
- I've gone with some repetition rather than DRY code.

Note that we don't load both JS files into the browser - we only load
the appropriate one based on the version of Wagtail in use